Crate tvm_sys

This crate contains the minimal interface over TVM’s C runtime API.

These common bindings are useful to both runtimes written in Rust, as well as higher level API bindings.

See the tvm-rt or tvm crates for full bindings to the TVM API.

Re-exports

pub use byte_array::ByteArray;

pub use context::Context;

pub use context::DeviceType;

pub use datatype::DataType;

pub use errors::*;

pub use packed_func::ArgValue;

pub use packed_func::RetValue;

Modules

array

byte_array

context
Provides Context and related device queries.

datatype

errors

ffi
The low-level C runtime FFI API for TVM.

packed_func

value

Macros

call_packed
Calls a packed function and returns a RetValue.

try_downcast