Module tvm_sys::context

Provides Context and related device queries.

Create a new context for device type and device id.

Example

let cpu = DeviceType::from("cpu");
let ctx = Context::new(cpu , 0);
let cpu0 = Context::cpu(0);
assert_eq!(ctx, cpu0);

Or from a supported device name.

use tvm_sys::Context;
let cpu0 = Context::from("cpu");
println!("{}", cpu0);

Structs

Context

UnsupportedDeviceError

Enums

DeviceType
Device type represents the set of devices supported by TVM.