Function tvm_sys::ffi::TVMDeviceCopyDataFromTo

pub unsafe extern "C" fn TVMDeviceCopyDataFromTo(
    from: *const c_void, 
    from_offset: usize, 
    to: *mut c_void, 
    to_offset: usize, 
    num_bytes: usize, 
    ctx_from: TVMContext, 
    ctx_to: TVMContext, 
    type_hint: DLDataType, 
    stream: TVMStreamHandle
) -> c_int

\brief Copy data from one place to another. \param from The source array. \param from_offset The byte offeset in the from. \param to The target array. \param to_offset The byte offset in the to. \param num_bytes The size of the memory in bytes \param ctx_from The source context \param ctx_to The target context \param type_hint The type of elements, only neded by certain backends. can be useful for cross device endian converison. \param stream Optional stream object. \return 0 when success, -1 when failure happens.