Crate tvm_sys

This crate contains the minimal interface over TVM’s C runtime API.

These common bindings are useful to both runtimes written in Rust, as well as higher level API bindings.

See the tvm-rt or tvm crates for full bindings to the TVM API.

Re-exports

pub use byte_array::ByteArray;

pub use datatype::DataType;

pub use device::Device;

pub use device::DeviceType;

pub use errors::*;

pub use packed_func::ArgValue;

pub use packed_func::RetValue;

Modules

array

byte_array

datatype

device

Provides Device and related device queries.

errors

ffi

The low-level C runtime FFI API for TVM.

packed_func

value

Macros

call_packed

Calls a packed function and returns a RetValue.

try_downcast